Improve recovery strategy missing pattern/expr/typexpr.
Instead of randomly aborting a parse when a part of the grammar might be
missing, try to skip nonsense tokens first. Based on the token that we
arrive on, a possible restart of the parsing might happen.

Example:

```
let x = = doStuff(y)
```

What we see here is an extra `=`, that doesn't make any sense in this
context. However if we just skip the `=` and restart the parse, we
gracefully recover. This is basically the aim of this PR.