Initial work on printing of braced expressions.
Context:

Imagine the user typing:
  let f = () => {
    a
  }

The printer will reformat this into:
  let f = () => a

This is very annoying. Inserting braces manually costs a lot of energy,
both mentally as physically when typing. To reduce friction, we preserve
the braces.

The printer uses the following logic:
// on same line
let f = (a, b) => {a + b}

// printer keeps it on one line
let f = (a, b) => { a + b }

// on different line
let f = (a, b) => {a + b
}
// or
let f = (a, b) => {
  a + b}

// printer keeps it on one line
let f = (a, b) => {
  a + b
}

Since the AST doesn't have a notion of a braced expression, an attribute
"ns.braces" is attached on expressions where braces are preserved.

There are currently two expressions where braces are preserved:
- rhs expr of a let-binding:
 let x = { a }
- return expression of an arrow expreesion:
 let f = () => { a }